refactor(section): move sections to folder
