+++ packages/validation/src/index.ts
@@ -22,7 +22,7 @@ export * from './modals/rejectApplicationModal.js';
 export * from './modals/requestRevisionsModal.js';
 export * from './routes/index.js';
 export * from './schemas.js';
-export * from './sections.js';
+export * from './section/sections.js';
 export * from './types.js';
 export * from './user.js';
 export * from './utils/functions.js';
